Add files via upload
Changes a number of default actions in 3rd party module. Forces all orders to appear as successful unless most basic parameters fail - employer wishes to call customer and secure the sale instead of hoping they'll continue to try if their credit card  gets a soft decline for the sale. Saves ALL payment methods as tokens, regardless of input, but makes them reusable by logged in customers when they opt-in (many customers do not log in to order, but call in to re-order, and they tend to prefer their payment methods being on file... IDK, I just build the stuff).